Refactor extension, add cohere and openai embeddings to SQL
