+++ lantern_extras/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lantern_extras"
-version = "0.0.10"
+version = "0.0.11"
 edition = "2021"
 
 [lib]
@@ -27,6 +27,7 @@ lantern_embeddings_core = { path = "../lantern_embeddings_core" }
 lantern_external_index = { path = "../lantern_external_index" }
 anyhow = "1.0.75"
 rand = "0.8.5"
+serde_json = "1.0.111"
 
 [dev-dependencies]
 pgrx-tests = "=0.9.7"
